deployment settings
