--- src/components/index.ts
@@ -1,1 +1,0 @@
-export { default as Roller } from "./Roller/Roller.vue";
