--- src/index.d.ts
@@ -1,3 +1,0 @@
-declare module "vue-roller" {
-    export default class Roller extends Vue {}
-}
